% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifequal.R
\name{EDagger}
\alias{EDagger}
\title{Total Life Expectancy Lost}
\usage{
EDagger(dx, exdagger, radix = 1)
}
\arguments{
\item{dx}{age distribution of deaths}

\item{exdagger}{life expectancy lost at age x}

\item{radix}{initial lifetable population (dx scaling factor)}
}
\description{
Life expectancy lost due to death.
}
\examples{
# total life expectancy lost for a 1x1 lifetable of Swedish females
swe      <- subset(sweden1x1, period == 1757 & sex == "female")
exdagger <- ExDagger(x = swe$x, ex = swe$ex)
EDagger(dx = swe$dx, exdagger = exdagger, radix = 100000)

# total life expectancy lost for a 5x5 lifetable of Swedish females
swe      <- subset(sweden5x5, period == "1755-1759" & sex == "female")
swe$wx[is.na(swe$wx)] <- 2 # last age group (110+) is assumed to be 2 years wide
exdagger <- ExDagger(x = swe$x, ex = swe$ex, wx = swe$wx, ax = swe$ax)
EDagger(dx = swe$dx, exdagger = exdagger, radix = 100000)

}

