% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifequal.R
\name{KeyfzEntro}
\alias{KeyfzEntro}
\title{Keyfitz's Entropy}
\usage{
KeyfzEntro(edagger, e0)
}
\arguments{
\item{edagger}{total life expectancy lost due to death}

\item{e0}{life expectancy at birth}
}
\description{
A measure for the lifespan equality in a population.
}
\examples{
  # lifespan equality for a 1x1 lifetable of Swedish females
  swe      <- subset(sweden1x1, period == 1757 & sex == "female")
  exdagger <- ExDagger(x = swe$x, ex = swe$ex)
  edagger  <- EDagger(dx = swe$dx, exdagger = exdagger, radix = 100000)
  KeyfzEntro(edagger = edagger, e0 = swe$ex[1])

  # lifespan equality for a 5x5 lifetable of Swedish females
  swe      <- subset(sweden5x5, period == "1755-1759" & sex == "female")
  exdagger <- ExDagger(x = swe$x, ex = swe$ex, wx = swe$wx, ax = swe$ax)
  edagger  <- EDagger(dx = swe$dx, exdagger = exdagger, radix = 100000)
  KeyfzEntro(edagger = edagger, e0 = swe$ex[1])

}

