% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifequal.R
\docType{package}
\name{lifequal}
\alias{lifequal}
\alias{lifequal-package}
\title{lifequal: Calculate Lifespan Equality from Life-tables}
\description{
\code{lifequal} lets you calculate three measures of lifespan equality from a
life-table:
}
\details{
1) \code{ExDagger(x, ex, ax)}
  Life expectancy lost by those who die in age interval [x, x+w)

2) \code{EDagger(dx, exdagger, radix)}
  Total life expectancy lost due to death

3) \code{KeyfzEntro(edagger, e0)} Keyfitz's entropy
}

