% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifequal.R
\name{ExDagger}
\alias{ExDagger}
\title{Life Expectancy Lost in Age x}
\usage{
ExDagger(x, ex, ax = 0.5)
}
\arguments{
\item{x}{start of age interval}

\item{ex}{life expectancy at age x}

\item{ax}{fraction of time spent in age interval x before dying in that
interval}
}
\description{
Life expectancy lost by those who die in age interval [x, x+w).
}
\details{
We assume that `x`, `ex` and `ax` are columns of a single, standard
  format lifetable: 1) ages are ordered from low to high, 2) there are no
  gaps between subsequent age intervals, 3) the last age interval is open to
  the right (e.g. 110+).

  The age interval widths are calculated automatically as the difference
  between the starting points of the age intervals `x`.

  For the last age interval omega we define:
  edagger(omega) = [e(omega) + 1.4] / 2
}
\examples{
# life expectancy lost in age x for a 1x1 lifetable of Swedish females
swe <- subset(sweden1x1, period == 1757 & sex == "female")
ExDagger(x = swe$x, ex = swe$ex)

# life expectancy lost in age [x,x+wx) for a 5x5 lifetable of Swedish females
swe <- subset(sweden5x5, period == "1755-1759" & sex == "female")
ExDagger(x = swe$x, ex = swe$ex, ax = swe$ax)

}

